application/app/requirements.txt
Flask==1.0.2
Keras==2.2.4
numpy==1.16.2
pandas==0.24.1
Pillow==5.1.0
scikit-learn==0.20.3
scipy==1.2.1
tensorflow==1.13.1
mysql-connector




